Ordered application files, moved all endpoint to views.py
