+++ app/models.py
@@ -9,4 +9,5 @@ class Product(db.Model):
     price = db.Column(db.Float())
 
     def __repr__(self):
-        return f'Product <id {self.id}><name {self.name}><price {self.price}'
+
+        return f'Name: {self.name}   Price: {self.price}   [ID:{self.id}]'
