+++ app/requirements.txt
@@ -1,0 +1,22 @@
+alembic==1.0.11
+autoenv==1.0.0
+Click==7.0
+Flask==1.0.3
+Flask-Migrate==1.8.0
+Flask-MySQLdb==0.2.0
+Flask-Script==2.0.6
+Flask-SQLAlchemy==2.1
+itsdangerous==1.1.0
+Jinja2==2.10.1
+Mako==1.0.12
+MarkupSafe==1.1.1
+mysql-connector==2.2.9
+mysqlclient==1.4.2.post1
+pkg-resources==0.0.0
+psycopg2==2.8.3
+PyMySQL==0.9.3
+python-dateutil==2.8.0
+python-editor==1.0.4
+six==1.12.0
+SQLAlchemy==1.3.5
+Werkzeug==0.15.4
